Update test runner (vstest.console)

+++ RunTest.ps1
@@ -1,18 +1,15 @@
 #!/usr/bin/env powershell -File
 
-param([string]$platform)
+param(
+  [string] $platform,
+  [switch] $clean
+)
 
 $projectname = "Wetzikon.Tests"
 $displayname = "Scrolliris Desktop for Windows (Tests)"
 $packagename = "Scrolliris.Desktop.Windows.Tests"
 $configuration = "Debug"
-$log = "Trace"
-
-Write-Host ""
-Write-Host "Configuration: $configuration"
-Write-Host "Platform: $platform"
-Write-Host "Log: $log"
-Write-Host ""
+$version = "1.0.0.0"
 
 # NOTE
 #
@@ -26,26 +23,38 @@ if ($platform -ne "x86" -and $platform -ne "x64" -and $platform -ne "ARM") {
   exit 1
 }
 
+Write-Host
+Write-Host "Configuration: $configuration"
+Write-Host "Platform: $platform"
+Write-Host "Clean: $clean"
+Write-Host
+
 taskkill /im 'MSBuild.exe' /f
 taskkill /im "testrunner.exe" /f
 
+$currentDir = (Get-Item ".\").FullName
+$packageDir = "${currentDir}\${projectname}" + `
+  "\AppPackages\${projectname}_${version}_${platform}_${configuration}_Test"
+$binDir = "${currentDir}\${projectname}\bin\${platform}\${configuration}"
 
-# Clean (resources and cache etc.)
-wsl rm -f `
-  "${projectname}/{bin,obj}/${platform}/${configuration}/${displayname}.exe"
-wsl rm -fr "${projectname}/{bin,obj}/${platform}/${configuration}/*"
+# Clean
+if ($clean) {
+  wsl rm -f `
+    "${projectname}/{bin,obj}/${platform}/${configuration}/${displayname}.exe"
+  wsl rm -fr "${projectname}/{bin,obj}/${platform}/${configuration}/*"
+  wsl rm -fr "${packageDir}/*"
 
-MSBuild.exe .\"${projectname}"\"${projectname}".csproj /t:Restore
-MSBuild.exe .\"${projectname}"\"${projectname}".csproj /t:Clean
+  Remove-Item "${packageDir}" -Recurse -Force
+
+  MSBuild.exe .\"${projectname}"\"${projectname}".csproj /t:Clean
+  MSBuild.exe .\"${projectname}"\"${projectname}".csproj /t:Restore
+}
 
 
 # Build
-$binDir = ".\${projectname}\bin\${platform}\${configuration}"
-
 MSBuild.exe .\"${projectname}"\"${projectname}".csproj /t:Build `
   /p:Configuration="${configuration}" `
-  /p:Platform="${platform}" `
-  /p:Log="${log}"
+  /p:Platform="${platform}"
 
 if ($lastexitcode -ne 0) {
   Write-Host "Build faild with status: ${lastexitcode}"
@@ -53,18 +62,28 @@ if ($lastexitcode -ne 0) {
 }
 
 
-# Run unit tests
-$origin = $PWD
-$location = Split-Path -Path $MyInvocation.MyCommand.Definition -Parent
+# Run tests with vstest
 
-Push-Location $location
-[Environment]::CurrentDirectory = $location
+# NOTE:
+# It seems that `vstest.console.exe` binaries in following directories won't
+# run correctly, appearently. Because it depends apparently VS :'(
 
-# NOTE: Install TestRunner via `NuGet.exe install Wetzikon.Tests/packages.config`
-$runnerVersion = "1.7.1"
-$runner = "$location\Packages\TestRunner.${runnerversion}\tools\testrunner.exe"
+$vsDir = "C:\Program Files (x86)\Microsoft Visual Studio\2017"
+#$toolsDir = "${vsDir}\TestAgent\Common7\IDE"
+#$toolsDir = "${vsDir}\BuildTools\Common7\IDE"
+$toolsDir = "${vsDir}\Community\Common7\IDE"
 
-& "$runner" "${binDir}\${displayname}.exe"
+# NOTE:
+# What the difference of these? ... :'(
 
-Pop-Location
-[Environment]::CurrentDirectory = $origin
+#$vstest = "${toolsDir}\CommonExtensions\Microsoft\TestWindow\vstest.console.exe"
+$vstest = "${toolsDir}\Extensions\TestPlatform\vstest.console.exe"
+
+Write-Host
+
+& "$vstest" `
+  /Platform:x64 `
+  /Framework:FrameworkUap10 `
+  /TestAdapterPath:"${projectname}\bin\${platform}\${configuration}" `
+  "${binDir}\${projectname}.build.appxrecipe" `
+  /Diag:VSTest.Console.log
